feat: rename modal folder to Modals
